requirements.txt
# Django Core
Django>=6.0,<7.0

# Environment configuration
python-decouple>=3.8,<4.0

# PostgreSQL support (for production)
psycopg2-binary>=2.9,<3.0

# Django REST Framework
djangorestframework>=3.15,<4.0
djangorestframework-simplejwt>=5.3,<6.0

# API Documentation
drf-spectacular>=0.27,<1.0

# Existing dependencies
django-simple-history>=3.4,<4.0
django-tailwind>=3.8,<4.0
django-browser-reload>=1.12,<2.0
Pillow>=11.0

# CORS (for React frontend)
django-cors-headers>=4.3,<5.0

# Filtering for DRF
django-filter>=24.0,<25.0

# Django Channels (WebSocket support)
daphne>=4.1,<5.0
channels>=4.0,<5.0
channels-redis>=4.2,<5.0

# PDF Generation
weasyprint>=62.0,<63.0
